addFarmerPrice

+++ app/src/main/java/com/farmers/underground/remote/models/UserPriceQualityModel.java
@@ -1,25 +1,26 @@
 package com.farmers.underground.remote.models;
 
 /**
- * Created by samson on 28.10.15.
+ * Created by samson
+ * on 28.10.15.
  */
 public class UserPriceQualityModel {
-    private float price;
+    private Double price;
     private String userQuality;
 
     public UserPriceQualityModel() {
     }
 
-    public UserPriceQualityModel(float price, String userQuality) {
+    public UserPriceQualityModel(Double price, String userQuality) {
         this.price = price;
         this.userQuality = userQuality;
     }
 
-    public float getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
